Fix imports in data_loader
